hw/ip/spatz: :bug: Clear MXU internal state upon new instruction
This is more of a hack to get the correct result and works
if the mxu results are accepted by the VRF as soon as they are ready.
If the VRF can also be non-ready sometimes, increase the delay
in the MXU counter by some amount. Everything works if the
number of cycles between the start of an MXU operation and its
final results is greater than the counter threshold to delay the
state clearing.