mxfp8-dotp: Initial working version
This is still flaky and the results seem to depend on the amount of debug
output, which still needs to be investigated.

Also, this runs on a single cluster core for the moment due to some issues with
the `snrt_cluster_hw_barrier()`.

The data file has been generated by hand for the moment.