mxfp8-matmul: Add mxfp8_matmul_fp32_rowmaj_m4
This kernel requires that both A and B matrices are stored in row-major format,
even though B is still block-quantized in the column direction. This allows an
implementation that does not use (expensive) reductions and is thus faster.
If B is (for example) a weight matrix that is prepared off-line, this constraint
should not be an issue.

Performance on (16,16,64): 22.8k cycles, 0.72 MAC/cycle