hw: Enable msip and jump to ISR on interrupt in bootrom
* In the current standalone tb, snitch cores moves out of the low-power
state (wfi) via debug requests. While this is fine for both standalone
and in-system behaviour, it is recommended that the sw stack supports
interrupt-based wake up. To do that, we modify the default exception
handler of the bootrom to point to `_snrt_isr` instead of an infinite
loop that does nothing.